+++ sql/core/src/main/scala/org/apache/spark/sql/execution/adaptive/AdaptiveSparkPlanExec.scala
@@ -117,7 +117,6 @@ case class AdaptiveSparkPlanExec(
       EnsureRequirements(requiredDistribution.isDefined, requiredDistribution)
     Seq(
       RemoveRedundantProjects,
-      OptimizeStoragePartitionedJoin,
       ensureRequirements,
       AdjustShuffleExchangePosition,
       ValidateSparkPlan,
