--- sql/core/src/main/scala/org/apache/spark/sql/execution/exchange/ReorderJoinKeys.scala
@@ -1,164 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more
- * contributor license agreements.  See the NOTICE file distributed with
- * this work for additional information regarding copyright ownership.
- * The ASF licenses this file to You under the Apache License, Version 2.0
- * (the "License"); you may not use this file except in compliance with
- * the License.  You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.spark.sql.execution.exchange
-
-import scala.collection.mutable
-import scala.collection.mutable.ArrayBuffer
-
-import org.apache.spark.sql.catalyst.expressions.Expression
-import org.apache.spark.sql.catalyst.plans.physical.HashPartitioning
-import org.apache.spark.sql.catalyst.plans.physical.KeyGroupedPartitioning
-import org.apache.spark.sql.catalyst.plans.physical.Partitioning
-import org.apache.spark.sql.catalyst.plans.physical.PartitioningCollection
-import org.apache.spark.sql.execution.SparkPlan
-import org.apache.spark.sql.execution.joins.ShuffledHashJoinExec
-import org.apache.spark.sql.execution.joins.SortMergeJoinExec
-
-/**
- * When the physical operators are created for JOIN, the ordering of join keys is based on order
- * in which the join keys appear in the user query. That might not match with the output
- * partitioning of the join node's children (thus leading to extra sort / shuffle being
- * introduced). This class is used to change the ordering of the join keys to match with the
- * partitioning of the join nodes' children.
- *
- * Note this is not a rule but rather an utility method used by [[EnsureRequirements]] and
- * [[OptimizeStoragePartitionedJoin]].
- */
-object ReorderJoinKeys {
-  def reorderJoinPredicates(plan: SparkPlan): SparkPlan = {
-    plan match {
-      case ShuffledHashJoinExec(
-      leftKeys, rightKeys, joinType, buildSide, condition, left, right, isSkew, isSPJ) =>
-        val (reorderedLeftKeys, reorderedRightKeys) =
-          reorderJoinKeys(leftKeys, rightKeys, left.outputPartitioning, right.outputPartitioning)
-        ShuffledHashJoinExec(reorderedLeftKeys, reorderedRightKeys, joinType, buildSide, condition,
-          left, right, isSkew, isSPJ)
-
-      case SortMergeJoinExec(leftKeys, rightKeys, joinType, condition, left, right, isSkew,
-          isSPJ) =>
-        val (reorderedLeftKeys, reorderedRightKeys) =
-          reorderJoinKeys(leftKeys, rightKeys, left.outputPartitioning, right.outputPartitioning)
-        SortMergeJoinExec(reorderedLeftKeys, reorderedRightKeys, joinType, condition,
-          left, right, isSkew, isSPJ)
-
-      case other => other
-    }
-  }
-
-  private def reorderJoinKeys(
-      leftKeys: Seq[Expression],
-      rightKeys: Seq[Expression],
-      leftPartitioning: Partitioning,
-      rightPartitioning: Partitioning): (Seq[Expression], Seq[Expression]) = {
-    if (leftKeys.forall(_.deterministic) && rightKeys.forall(_.deterministic)) {
-      reorderJoinKeysRecursively(
-        leftKeys,
-        rightKeys,
-        Some(leftPartitioning),
-        Some(rightPartitioning))
-          .getOrElse((leftKeys, rightKeys))
-    } else {
-      (leftKeys, rightKeys)
-    }
-  }
-
-  /**
-   * Recursively reorders the join keys based on partitioning. It starts reordering the
-   * join keys to match HashPartitioning on either side, followed by PartitioningCollection.
-   */
-  private def reorderJoinKeysRecursively(
-      leftKeys: Seq[Expression],
-      rightKeys: Seq[Expression],
-      leftPartitioning: Option[Partitioning],
-      rightPartitioning: Option[Partitioning]): Option[(Seq[Expression], Seq[Expression])] = {
-    (leftPartitioning, rightPartitioning) match {
-      case (Some(HashPartitioning(leftExpressions, _)), _) =>
-        reorder(leftKeys.toIndexedSeq, rightKeys.toIndexedSeq, leftExpressions, leftKeys)
-            .orElse(reorderJoinKeysRecursively(
-              leftKeys, rightKeys, None, rightPartitioning))
-      case (_, Some(HashPartitioning(rightExpressions, _))) =>
-        reorder(leftKeys.toIndexedSeq, rightKeys.toIndexedSeq, rightExpressions, rightKeys)
-            .orElse(reorderJoinKeysRecursively(
-              leftKeys, rightKeys, leftPartitioning, None))
-      case (Some(KeyGroupedPartitioning(clustering, _, _)), _) =>
-        val leafExprs = clustering.flatMap(_.collectLeaves())
-        reorder(leftKeys.toIndexedSeq, rightKeys.toIndexedSeq, leafExprs, leftKeys)
-            .orElse(reorderJoinKeysRecursively(
-              leftKeys, rightKeys, None, rightPartitioning))
-      case (_, Some(KeyGroupedPartitioning(clustering, _, _))) =>
-        val leafExprs = clustering.flatMap(_.collectLeaves())
-        reorder(leftKeys.toIndexedSeq, rightKeys.toIndexedSeq, leafExprs, rightKeys)
-            .orElse(reorderJoinKeysRecursively(
-              leftKeys, rightKeys, leftPartitioning, None))
-      case (Some(PartitioningCollection(partitionings)), _) =>
-        partitionings.foldLeft(Option.empty[(Seq[Expression], Seq[Expression])]) { (res, p) =>
-          res.orElse(reorderJoinKeysRecursively(leftKeys, rightKeys, Some(p), rightPartitioning))
-        }.orElse(reorderJoinKeysRecursively(leftKeys, rightKeys, None, rightPartitioning))
-      case (_, Some(PartitioningCollection(partitionings))) =>
-        partitionings.foldLeft(Option.empty[(Seq[Expression], Seq[Expression])]) { (res, p) =>
-          res.orElse(reorderJoinKeysRecursively(leftKeys, rightKeys, leftPartitioning, Some(p)))
-        }.orElse(None)
-      case _ =>
-        None
-    }
-  }
-
-  private def reorder(
-      leftKeys: IndexedSeq[Expression],
-      rightKeys: IndexedSeq[Expression],
-      expectedOrderOfKeys: Seq[Expression],
-      currentOrderOfKeys: Seq[Expression]): Option[(Seq[Expression], Seq[Expression])] = {
-    if (expectedOrderOfKeys.size != currentOrderOfKeys.size) {
-      return None
-    }
-
-    // Check if the current order already satisfies the expected order.
-    if (expectedOrderOfKeys.zip(currentOrderOfKeys).forall(p => p._1.semanticEquals(p._2))) {
-      return Some(leftKeys, rightKeys)
-    }
-
-    // Build a lookup between an expression and the positions its holds in the current key seq.
-    val keyToIndexMap = mutable.Map.empty[Expression, mutable.BitSet]
-    currentOrderOfKeys.zipWithIndex.foreach {
-      case (key, index) =>
-        keyToIndexMap.getOrElseUpdate(key.canonicalized, mutable.BitSet.empty).add(index)
-    }
-
-    // Reorder the keys.
-    val leftKeysBuffer = new ArrayBuffer[Expression](leftKeys.size)
-    val rightKeysBuffer = new ArrayBuffer[Expression](rightKeys.size)
-    val iterator = expectedOrderOfKeys.iterator
-    while (iterator.hasNext) {
-      // Lookup the current index of this key.
-      keyToIndexMap.get(iterator.next().canonicalized) match {
-        case Some(indices) if indices.nonEmpty =>
-          // Take the first available index from the map.
-          val index = indices.firstKey
-          indices.remove(index)
-
-          // Add the keys for that index to the reordered keys.
-          leftKeysBuffer += leftKeys(index)
-          rightKeysBuffer += rightKeys(index)
-        case _ =>
-          // The expression cannot be found, or we have exhausted all indices for that expression.
-          return None
-      }
-    }
-    Some(leftKeysBuffer.toSeq, rightKeysBuffer.toSeq)
-  }
-}
